+++ src/main/java/org/ojalgo/function/ComplexFunction.java
@@ -36,6 +36,10 @@ public final class ComplexFunction extends FunctionSet<ComplexNumber> {
             return this.invoke(ComplexNumber.valueOf(arg1), ComplexNumber.valueOf(arg2)).doubleValue();
         }
 
+        default ComplexNumber invoke(final ComplexNumber arg1, final double arg2) {
+            return this.invoke(arg1, ComplexNumber.valueOf(arg2));
+        }
+
         default float invoke(final float arg1, final float arg2) {
             return this.invoke(ComplexNumber.valueOf(arg1), ComplexNumber.valueOf(arg2)).floatValue();
         }