+++ CHANGELOG.md
@@ -17,6 +17,16 @@ Added / Changed / Deprecated / Fixed / Removed / Security
 
 - There was a case when the `IntegerSolver` returned an incorrect solution (constraint breaking) but reported it to be optimal. It was actually the `LinearSolver` that malfunctioned, but behaviour in the `IntegerSolver` that generated the problematic node model. Fixed this problem by 1) making sure the `LinearSolver` handles that case, and 2) altered the problematic behaviour in the `IntegerSolver` to be "safer".
 
+### Changed
+
+#### org.ojalgo.array
+
+- More efficient implementation of `reduce(int,int,Aggregator)` in `ArrayAnyD` (reduce to a 2D structure).
+
+#### org.ojalgo.structure
+
+- Deprecated `loop(Predicate<long[]>,IndexCallback)` in favour of the new `loopReferences(Predicate<long[]>,ReferenceCallback)` method in `StructureAnyD`.
+
 ## [53.0.0] – 2023-04-16
 
 ### Added